主子表代码生成优化

+++ ox-admin-application/src/main/java/com/java668/oxadmin/test/dto/request/SystUserReqDTO.java
@@ -53,5 +53,5 @@ public class SystUserReqDTO {
     /**
      * 是否删除（0:否，1：是）
      */
-    private Long isDeleted;
+    private Integer isDeleted;
 }
