+++ ox-admin-application/src/main/java/com/java668/oxadmin/test/dto/response/SystUserRespDTO.java
@@ -54,5 +54,5 @@ public class SystUserRespDTO extends BaseDTO {
     /**
      * 是否删除（0:否，1：是）
      */
-    private Long isDeleted;
+    private Integer isDeleted;
 }
